requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchainhub==0.1.15
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchain-google-community==1.0.2
openai==1.23.2
tiktoken==0.6.0
deeplake==3.9.1
google-api-python-client==2.127.0
python-dotenv==1.0.1
mypy==1.9.0
